// lib/components/Text/types.ts
export type CoreTextVariant =
  | 'heading-72'
  | 'heading-64'
  | 'heading-56'
  | 'heading-48'
  | 'heading-40'
  | 'heading-32'
  | 'heading-24'
  | 'heading-20'
  | 'heading-16'
  | 'heading-14'
  | 'copy-24'
  | 'copy-20'
  | 'copy-18'
  | 'copy-16'
  | 'copy-14'
  | 'copy-13';

export type Size =
  | 48
  | 40
  | 32
  | 24
  | 20
  | 18
  | 16
  | 14
  | 13
  | 12
  | 10;


export type Weight =
  | 700
  | 600
  | 500
  | 400;

export type LineHeight = number